backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pgvector==0.2.5
httpx==0.27.0
openai>=1.58.1,<2.0.0
tiktoken>=0.7.0,<1.0.0
websockets==12.0
langchain>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0
langchain-groq>=0.2.0,<0.3.0
langchain-ollama>=0.3.0,<0.4.0
langgraph>=0.2.0,<0.3.0

